FEAT: Copy only when different
No need to reload picom for other app's config updates again!

+++ copy_to_config.sh
@@ -5,17 +5,22 @@ config_folder=~/.config/
 mkdir -p "$config_folder"
 
 # For every file in this folder
-for file in $(find -type f | \
-    sed "/copy_to_config.*/d" | sed "/README.*/d" | \
-    sed "/tips\//d" | sed "/\.git\//d")
+for file in $(find -type f | sed "/copy_to_config.*/d" | \
+    sed "/README.*/d" | sed "/tips\//d" | sed "/\.git\//d")
 do
-    # Copy the file over, creating parent folders if necessary
-    cp --parents "$file" "$config_folder" 2>/dev/null
+    original_file="${config_folder}/${file}"
+
+    # Check if the file is different
+    if ! test -f "$original_file" || ! cmp -s "$file" "$original_file"
+    then
+        # Copy the file over, creating parent folders if necessary
+        cp --parents "$file" "$config_folder" 2>/dev/null
+    fi
 
     # Make necessary files executable
     mod=$(stat -c "%a" "$file" | cut -b 1)
     if ((mod % 2 == 1))
     then
-        chmod +x "$config_folder"/"$file"
+        chmod +x "$original_file"
     fi
 done
